+++ pp/components/delay_snake3.py
@@ -4,7 +4,7 @@ import numpy as np
 
 import pp
 from pp.component import Component
-from pp.tech import FACTORY, Factory
+from pp.tech import LIBRARY, Library
 from pp.types import StrOrDict, StrOrDictOrNone
 
 
@@ -15,7 +15,7 @@ def delay_snake3(
     n: int = 2,
     bend180: StrOrDictOrNone = None,
     waveguide: StrOrDict = "strip",
-    factory: Factory = FACTORY,
+    library: Library = LIBRARY,
     **waveguide_settings,
 ) -> Component:
     r"""Snake input facing west
@@ -27,7 +27,7 @@ def delay_snake3(
         n: number of loops
         bend180
         waveguide
-        factory
+        library
         waveguide_settings
 
     .. code::
@@ -51,7 +51,7 @@ def delay_snake3(
     bend180 = bend180 or dict(
         component="bend_euler", angle=180, waveguide=waveguide, **waveguide_settings
     )
-    bend180 = factory.get_component(bend180, waveguide=waveguide, **waveguide_settings)
+    bend180 = library.get_component(bend180, waveguide=waveguide, **waveguide_settings)
 
     delta_length = (length - length0 - n * (bend180.length)) / (n + 1)
     assert (