+++ pp/types.py
@@ -1,4 +1,4 @@
-"""In programming, a factory is a function that returns an Object.
+"""In programming, a library is a function that returns an Object.
 
 Functions are easy to understand because they have clear inputs and outputs.
 Most gdsfactory functions take some inputs and return a Component object.